Moved the tests from controller_utils_test.go to controller_test.go, and fixed the import statements
